shout_detection/requirements.txt
pandas
numpy
librosa
scipy
praat-parselmouth
matplotlib
tensorflow==1.15.5